+++ R/empi.install.R
@@ -1,10 +1,16 @@
-#' Installs the required external program
+#' Installs the EMPI external program
 #'
-#' Downloads \strong{Enhanced Matching Pursuit Implementation} external program (or EMPI for short)
-#' and stores it in the cache directory.
+#' Downloads \strong{Enhanced Matching Pursuit Implementation} external program
+#' (or EMPI for short) binary compatible with the current operating system and
+#' stores it in the package cache directory.
 #'
-#' @return The function downloads the EMPI program in a version compatible with the operating
-#' system used (Windows, Linux, MacOS-x64, MacOS-arm64) and stores it in the cache directory.
+#' The function detects the operating system (Windows, Linux, macOS x64,
+#' macOS arm64), downloads the appropriate archive from the official
+#' repository, verifies its integrity using a checksum, and extracts it.
+#'
+#' @return The function downloads the EMPI program in a version compatible
+#' with the operating system used (Windows, Linux, MacOS-x64, MacOS-arm64)
+#' and stores it in the package cache directory.
 #'
 #' @importFrom utils download.file unzip
 #'
@@ -26,22 +32,34 @@ empi.install <- function() {
   sys <- Sys.info()[["sysname"]]
   mach <- Sys.info()[["machine"]]
 
-  files <- list.files(dest.dir)
-  if (sys == "Windows") {
-    if ("empi.exe" %in% files) {
-      message("It looks like EMPI is already installed in the '", dest.dir, "' directory.")
-      return(invisible(NULL))
-    }
+  bin.exists <- if (sys == "Windows") {
+    file.exists(file.path(dest.dir, "empi.exe"))
   } else {
-    if ("empi" %in% files) {
-      message("It looks like EMPI is already installed in the '", dest.dir, "' directory.")
-      return(invisible(NULL))
-    }
+    file.exists(file.path(dest.dir, "empi"))
+  }
+
+  if (bin.exists) {
+    message("EMPI already installed in: '", dest.dir, "' directory.")
+    return(invisible(dest.dir))
   }
 
   message("Downloading EMPI (third-party software) for ", sys, " ", mach, "...")
 
-  download.file(url = out$url,  destfile = archive, mode = "wb")
+  ok <- tryCatch({
+    download.file(url = out$url, destfile = archive, mode = "wb", quiet = TRUE)
+    TRUE
+  }, error = function(e) {
+        message("Download failed: ", e$message)
+        FALSE
+    }
+  )
+
+  if (!ok) stop("Installation aborted due to download failure.")
+
+  if (!file.exists(archive)) {
+    stop("Downloaded archive not found.")
+  }
+
   check.checksum(archive)
 
   message("Extracting EMPI to '", dest.dir, "' ...")
@@ -53,7 +71,9 @@ empi.install <- function() {
     Sys.chmod(exec, "0755")
   }
 
-  message("Installation complete. EMPI program is in '", dest.dir, "'")
+  message("Installation complete. EMPI program is in '", dest.dir, "' directory.")
+
+  invisible(dest.dir)
 
 }